requirements.txt
# ─── Core RDF / SPARQL ────────────────────────────────────────────────────────
rdflib==7.1.1
SPARQLWrapper==2.0.0
pyshacl==0.27.0

# ─── Neo4j ────────────────────────────────────────────────────────────────────
neo4j==5.26.0

# ─── HTTP / Data ──────────────────────────────────────────────────────────────
requests==2.32.3
httpx==0.28.1
pandas==2.2.3
faker==33.1.0
tqdm==4.67.1
python-dotenv==1.0.1
pydantic==2.10.3

# ─── LangChain / LangGraph (latest as of Apr 2026) ────────────────────────────
langchain==0.3.27
langchain-core==0.3.61
langchain-community==0.3.27
langchain-neo4j==0.4.0
langchain-anthropic==0.3.3
langchain-openai==0.3.0
langgraph==0.2.70
langgraph-checkpoint==2.0.10

# ─── Neo4j GraphRAG (official Neo4j library, used in Going Meta S22-S45) ──────
neo4j-graphrag==1.7.0

# ─── Dashboard ────────────────────────────────────────────────────────────────
streamlit==1.41.1
plotly==5.24.1

# ─── Notebooks ────────────────────────────────────────────────────────────────
jupyter==1.1.1
ipykernel==6.29.5

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
